+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+* Updates README via code excerpts.
 * Fixes violations of new analysis option use_named_constants.
 
 ## 2.4.6